add PointsToFlowSensitiveWithoutMerge
This is a new experimental type of pointer analysis. It is flow-sensitive,
but instead of merging maps in joins, it does lookups backward
in the graph, so that we do not merge so much information.
This is just few steps from sparse PTA as if this lookup
was a bit smarter, we would almost have the sparse analysis.